Jump animation working
